Implemented bulk edit and remove for values in ResourceView; Added a warning message when removing a user from a project in the Project-Users management panel; Now when the user changes the project from the project-list modal, it is redirected to the home page.

+++ src/app/administration/projectsAdministration/projectUsersManagerComponent.ts
@@ -129,17 +129,22 @@ export class ProjectUsersManagerComponent {
     }
 
     private removeUserFromProject() {
-        this.adminService.removeUserFromProject(this.project.getName(), this.selectedUser.getEmail()).subscribe(
-            stResp => {
-                for (var i = 0; i < this.usersBound.length; i++) {
-                    if (this.usersBound[i].getEmail() == this.selectedUser.getEmail()) {
-                        this.usersBound.splice(i, 1);
+        this.basicModals.confirm("Remove user", "You are removing the user " + this.selectedUser.getShow() + " form the project " + this.project.getName() + ". Are you sure?", "warning").then(
+            () => {
+                this.adminService.removeUserFromProject(this.project.getName(), this.selectedUser.getEmail()).subscribe(
+                    stResp => {
+                        for (var i = 0; i < this.usersBound.length; i++) {
+                            if (this.usersBound[i].getEmail() == this.selectedUser.getEmail()) {
+                                this.usersBound.splice(i, 1);
+                            }
+                        }
+                        this.puBinding.setRoles([]);
+                        this.selectedUser = null;
+                        this.selectedRole = null;
                     }
-                }
-                this.puBinding.setRoles([]);
-                this.selectedUser = null;
-                this.selectedRole = null;
-            }
+                );
+            }, 
+            () => {}
         );
     }
